@@ -23,18 +23,23 @@
 
 package net.whimxiqal.journey.bukkit;
 
+import java.util.Arrays;
+import java.util.stream.Collectors;
 import net.whimxiqal.journey.Journey;
 import net.whimxiqal.journey.ProxyImpl;
+import net.whimxiqal.journey.bukkit.listener.PluginDisableListener;
 import net.whimxiqal.journey.command.JourneyConnectorProvider;
-import net.whimxiqal.journey.bukkit.config.BukkitConfigManager;
 import net.whimxiqal.journey.bukkit.listener.DeathListener;
 import net.whimxiqal.journey.bukkit.listener.NetherListener;
-import net.whimxiqal.journey.bukkit.search.listener.PlayerListener;
+import net.whimxiqal.journey.bukkit.listener.PlayerListener;
 import net.whimxiqal.journey.bukkit.util.BukkitLogger;
 import net.whimxiqal.journey.bukkit.util.BukkitSchedulingManager;
+import net.whimxiqal.journey.config.Settings;
+import net.whimxiqal.journey.util.Request;
 import net.whimxiqal.mantle.paper.PaperRegistrarProvider;
 import net.whimxiqal.mantle.common.CommandRegistrar;
 import org.bukkit.Bukkit;
+import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.java.JavaPlugin;
 
 public final class JourneyBukkit extends JavaPlugin {
@@ -73,7 +78,7 @@ public final class JourneyBukkit extends JavaPlugin {
     proxy.logger(new BukkitLogger());
     proxy.dataFolder(this.getDataFolder().toPath());
     proxy.audienceProvider(new PaperAudiences());
-    proxy.configManager(BukkitConfigManager.initialize("config.yml"));
+    proxy.configPath(this.getDataFolder().toPath().resolve("config.yml"));
     proxy.schedulingManager(new BukkitSchedulingManager());
     proxy.platform(new BukkitPlatformProxy());
     proxy.version(getDescription().getVersion());
@@ -102,6 +107,18 @@ public final class JourneyBukkit extends JavaPlugin {
     Bukkit.getPluginManager().registerEvents(new NetherListener(), this);
     Bukkit.getPluginManager().registerEvents(new PlayerListener(), this);
     Bukkit.getPluginManager().registerEvents(new DeathListener(), this);
+    Bukkit.getPluginManager().registerEvents(new PluginDisableListener(), this);
+
+    if (Settings.EXTRA_CHECK_LATEST_VERSION_ON_STARTUP.getValue()) {
+      Request.evaluateVersionAge("paper", getDescription().getVersion());
+    }
+    if (Settings.EXTRA_FIND_INTEGRATIONS_ON_STARTUP.getValue()) {
+      Request.checkForIntegrationPlugins("paper",
+          Bukkit.getMinecraftVersion(),
+          Arrays.stream(Bukkit.getPluginManager().getPlugins())
+              .map(Plugin::getName)
+              .collect(Collectors.toSet()));
+    }
   }
 
   @Override
