@@ -23,29 +23,10 @@
 
 package net.whimxiqal.journey.math;
 
-public class Vector {
-  private final double x;
-  private final double y;
-  private final double z;
-  
-  public Vector(double x, double y, double z) {
-    this.x = x;
-    this.y = y;
-    this.z = z;
-  }
+import java.util.Objects;
 
-  public double x() {
-    return x;
-  }
+public record Vector(double x, double y, double z) {
 
-  public double y() {
-    return y;
-  }
-
-  public double z() {
-    return z;
-  }
-  
   public Vector subtract(Vector other) {
     return new Vector(x - other.x, y - other.y, z - other.z);
   }
@@ -60,6 +41,12 @@ public class Vector {
 
   public double dot(Vector other) {
     return x * other.x + y * other.y + z * other.z;
+  }
+
+  public Vector cross(Vector other) {
+    return new Vector(this.y * other.z - this.z * other.y,
+        this.z * other.x - this.x * other.z,
+        this.x * other.y - this.y * other.x);
   }
 
   public double magnitude() {
@@ -78,4 +65,13 @@ public class Vector {
   public String toString() {
     return "Vector{" + "x=" + x + ", y=" + y + ", z=" + z + '}';
   }
+
+  @Override
+  public boolean equals(Object o) {
+    if (this == o) return true;
+    if (o == null || getClass() != o.getClass()) return false;
+    Vector vector = (Vector) o;
+    return Double.compare(vector.x, x) == 0 && Double.compare(vector.y, y) == 0 && Double.compare(vector.z, z) == 0;
+  }
+
 }
